refactor: split double_buffer module
